+++ src/test/core/WidgetTests.scala
@@ -3,6 +3,7 @@
 package org.nlogo.core
 
 import org.scalatest.FunSuite
+import ConstraintSpecification.ChoiceConstraintSpecification
 
 class WidgetTests extends FunSuite {
   test("Chooser handles choices with lists") {
@@ -11,7 +12,7 @@ class WidgetTests extends FunSuite {
       override def toString: String = toList.toString()
     }
 
-    val testChoices = List(l, 4.toDouble).asInstanceOf[List[AnyRef]]
+    val testChoices: List[AnyRef] = List(l, 4.toDouble.asInstanceOf[AnyRef])
     val chooser = Chooser(display = "FOOBAR", varName = "FOOBAR", choices = testChoices)
     assertResult(ChoiceConstraintSpecification(testChoices, 0))(chooser.constraint)
   }
